Move UpdateNodeSwitchExcludeIPs to libovsdb/util pkg
Removes dependency of util pkg on libovsdbops.
Ideally util should not import any other pkg so that
other packages can import the util functions.
